updated telegram chat id cols to big int and updated alembic functionality to check type

+++ app/models.py
@@ -11,7 +11,7 @@ class Subscribers(db.Model):
     platform = db.Column(db.String(60))
     number_support_texts = db.Column(db.Integer)
     timestamp = db.Column(db.DateTime, index=True, default=datetime.utcnow)
-    telegram_chat_id = db.Column(db.Integer)
+    telegram_chat_id = db.Column(db.BigInteger)
 
     def __repr__(self):
         return f"<id {self.id}>"
@@ -22,7 +22,7 @@ class MJMAlerts(db.Model):
     mjm_alerts = db.Column(db.Boolean)
     platform = db.Column(db.String(60))
     timestamp = db.Column(db.DateTime, index=True, default=datetime.utcnow)
-    telegram_chat_id = db.Column(db.Integer)
+    telegram_chat_id = db.Column(db.BigInteger)
 
     def __repr__(self):
         return f"<id {self.id}>"
